Refactor 'toPosition'
