+++ alchemist/alchemist-smartcam/src/main/kotlin/it/unibo/alchemist/model/implementations/actions/HeadTowardTarget.kt
@@ -1,17 +1,17 @@
 package it.unibo.alchemist.model.implementations.actions
 
+import it.unibo.alchemist.kotlin.toPosition
 import it.unibo.alchemist.model.interfaces.Context
 import it.unibo.alchemist.model.interfaces.Molecule
 import it.unibo.alchemist.model.interfaces.Node
 import it.unibo.alchemist.model.interfaces.Reaction
 import it.unibo.alchemist.model.interfaces.environments.Physics2DEnvironment
-import it.unibo.alchemist.model.smartcam.toPosition
+import org.apache.commons.math3.util.FastMath.toRadians
 import kotlin.math.abs
 import kotlin.math.cos
 import kotlin.math.min
 import kotlin.math.sign
 import kotlin.math.sin
-import org.apache.commons.math3.util.FastMath.toRadians
 
 /**
  * Reads the target's absolute coordinates from the [target] molecule
@@ -19,7 +19,7 @@ import org.apache.commons.math3.util.FastMath.toRadians
  */
 class HeadTowardTarget<T> @JvmOverloads constructor(
     node: Node<T>,
-    private val env: Physics2DEnvironment<T>,
+    private val environment: Physics2DEnvironment<T>,
     private val reaction: Reaction<T>,
     private val target: Molecule,
     private val angularSpeedDegrees: Double = 360.0
@@ -27,7 +27,8 @@ class HeadTowardTarget<T> @JvmOverloads constructor(
 
     private val angularSpeedRadians = toRadians(angularSpeedDegrees)
 
-    override fun cloneAction(n: Node<T>, r: Reaction<T>) = HeadTowardTarget(n, env, r, target, angularSpeedDegrees)
+    override fun cloneAction(n: Node<T>, r: Reaction<T>) =
+        HeadTowardTarget(n, environment, r, target, angularSpeedDegrees)
 
     /**
      * Sets the heading of the node according to the target molecule.
@@ -35,19 +36,19 @@ class HeadTowardTarget<T> @JvmOverloads constructor(
     override fun execute() {
         node.getConcentration(target)?.also {
             val speedRadians = angularSpeedRadians / reaction.timeDistribution.rate
-            val targetPosition = it.toPosition(env)
-            val myHeading = env.getHeading(node)
+            val targetPosition = it.toPosition(environment)
+            val myHeading = environment.getHeading(node)
             if (targetPosition != myHeading) {
                 if (speedRadians >= 2 * Math.PI) {
-                    env.setHeading(node, targetPosition - env.getPosition(node))
+                    environment.setHeading(node, targetPosition - environment.getPosition(node))
                 } else {
-                    val targetAngle = (targetPosition - env.getPosition(node)).asAngle
-                    val currentAngle = env.getHeading(node).asAngle
+                    val targetAngle = (targetPosition - environment.getPosition(node)).asAngle
+                    val currentAngle = environment.getHeading(node).asAngle
                     val rotation = shortestRotationAngle(currentAngle, targetAngle)
                     val absDistance = abs(rotation)
                     if (absDistance > 0) {
                         val newAngle = currentAngle + min(speedRadians, absDistance) * rotation.sign
-                        env.setHeading(node, env.makePosition(cos(newAngle), sin(newAngle)))
+                        environment.setHeading(node, environment.makePosition(cos(newAngle), sin(newAngle)))
                     }
                 }
             }
