+++ alchemist/alchemist-smartcam/src/main/kotlin/it/unibo/alchemist/model/smartcam/Utils.kt
@@ -2,23 +2,8 @@ package it.unibo.alchemist.model.smartcam
 
 import it.unibo.alchemist.model.implementations.positions.Euclidean2DPosition
 import it.unibo.alchemist.model.interfaces.Environment
-import it.unibo.alchemist.model.interfaces.Position
 import kotlin.math.cos
 import kotlin.math.sin
-
-internal inline fun <reified P : Position<P>> Any?.toPosition(env: Environment<*, P>): P = when (this) {
-    is P -> this
-    is Iterable<*> -> env.makePosition(*(
-        this.map {
-            when (it) {
-                is Number -> it
-                else -> throw IllegalStateException(
-                    "The Iterable must contain only Numbers but ${it?.javaClass} has been found"
-                )
-            }
-        }).toTypedArray())
-    else -> throw IllegalArgumentException("Expected an Iterable or Euclidean2DPosition but got a ${this?.javaClass}")
-}
 
 internal fun offsetPositionAtDistance(
     env: Environment<*, Euclidean2DPosition>,
